% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eafit.R
\name{theme_eafit_dark}
\alias{theme_eafit_dark}
\title{EAFIT dark theme}
\usage{
theme_eafit_dark(
  text.font = "Literation Mono Powerline",
  title.font = "Literation Mono Powerline",
  legend.font = "Literation Mono Powerline",
  title.size = 18,
  text.size = 14,
  subtitle.size = 12,
  axis.title.size = 14,
  axis.text.size = 12,
  legend.title.size = 10,
  legend.text.size = 9,
  title.color = "#FFFFFF",
  subtitle.color = "#FFFFFF",
  text.color = "#FFFFFF",
  axis.title.color = "#FFFFFF",
  axis.text.color = "#FFFFFF",
  legend.title.color = "#FFFFFF",
  legend.text.color = "#FFFFFF",
  legend.position = "bottom",
  ticks = FALSE
)
}
\description{
Darker version
}
