% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eafit.R
\name{theme_eafit_light}
\alias{theme_eafit_light}
\title{EAFIT light theme}
\usage{
theme_eafit_light(
  text.font = "Literation Mono Powerline",
  title.font = "Literation Mono Powerline",
  legend.font = "Literation Mono Powerline",
  title.size = 18,
  text.size = 14,
  subtitle.size = 12,
  axis.title.size = 14,
  axis.text.size = 12,
  legend.title.size = 10,
  legend.text.size = 9,
  title.color = "#000359",
  subtitle.color = "#000359",
  text.color = "#000359",
  axis.title.color = "#000359",
  axis.text.color = "#000359",
  legend.title.color = "#000359",
  legend.text.color = "#000359",
  legend.position = "bottom",
  ticks = FALSE
)
}
\description{
ligther
}
