% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eafit.R
\name{theme_eafit_light}
\alias{theme_eafit_light}
\title{EAFIT light theme}
\usage{
theme_eafit_light(
  text.font = NULL,
  title.font = NULL,
  legend.font = NULL,
  title.size = 18,
  text.size = 14,
  subtitle.size = 12,
  axis.title.size = 14,
  axis.text.size = 12,
  legend.title.size = 10,
  legend.text.size = 9,
  title.color = "#000359",
  subtitle.color = "#000359",
  text.color = "#000359",
  axis.title.color = "#000359",
  axis.text.color = "#000359",
  legend.title.color = "#000359",
  legend.text.color = "#000359",
  legend.position = "bottom"
)
}
\arguments{
\item{text.font}{font, Default: NULL}

\item{title.font}{font, Default: NULL}

\item{legend.font}{font, Default: NULL}

\item{title.size}{title font size, Default: 18}

\item{text.size}{text font size, Default: 14}

\item{subtitle.size}{subtitle font size, Default: 12}

\item{axis.title.size}{title font size, Default: 14}

\item{axis.text.size}{text font size, Default: 12}

\item{legend.title.size}{title font size, Default: 10}

\item{legend.text.size}{text font size, Default: 9}

\item{title.color}{title color, Default: '#000359'}

\item{subtitle.color}{subtitle color, Default: '#000359'}

\item{text.color}{text color, Default: '#000359'}

\item{axis.title.color}{axis title color, Default: '#000359'}

\item{axis.text.color}{axis text color, Default: '#000359'}

\item{legend.title.color}{legend title color, Default: '#000359'}

\item{legend.text.color}{legend text color, Default: '#000359'}

\item{legend.position}{legend position, Default: 'bottom'}
}
\description{
ligther
}
\examples{
library(ggplot2)
library(palmerpenguins)
library(eafithemer)

ggplot(penguins, aes(bill_length_mm, flipper_length_mm)) +
  geom_point(aes(color = species, shape = species)) +
  labs(
    title = "Little title",
    subtitle = "Testing the subtitle",
    caption = "First attempt on package creation"
  ) +
  eafithemer::theme_eafit_light() +
  eafithemer::scale_color_eafit() +
  theme(
    legend.title = element_blank()
  )
}
